pyproject.toml
[tool.poetry]
name = "etl-github"
version = "0.1.0"
description = ""
authors = ["none"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jupyter = "^1.0.0"
ipykernel = "^6.29.4"
pandas = "^2.2.2"
black = "^24.4.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
